Use gsize for length of data fields, not int; use unsigned int insted of
2007-05-11  Loïc Minier  <lool@dooz.org>

       * libebook/e-contact.h:
       Use gsize for length of data fields, not int; use unsigned int insted
       of int for year, month, and day. (#414191)

svn path=/trunk/; revision=7737

+++ addressbook/libebook/e-contact.h
@@ -232,7 +232,7 @@ typedef struct {
 	union {
 		struct {
 			char *mime_type;
-			int length;
+			gsize length;
 			guchar *data;
 		} inlined;
 		char *uri;
@@ -253,13 +253,13 @@ typedef struct {
 } EContactAddress;
 
 typedef struct {
-	int year;
-	int month;
-	int day;
+	unsigned int year;
+	unsigned int month;
+	unsigned int day;
 } EContactDate;
 
 typedef struct {
-	int length;
+	gsize length;
 	char *data;
 } EContactCert;
